Setup simple stack coordinator
